backend/pyproject.toml
[tool.poetry]
name = "fastapi-docker-compose-backend"
version = "0.1.0"
description = "A backend for a demo fastapi fullstack project"
authors = ["Sam McVicker <sammcvicker@protonmail.com>"]
license = "MIT"
readme = "README.md"
include = ["src"]

[tool.poetry.dependencies]
python = "^3.13"
asyncpg = "^0.30.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
python-multipart = "^0.0.19"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
